requirements.txt
accelerate==0.25.0
decord==0.6.0
einops==0.7.0
fire==0.5.0
human_body_prior==0.8.5.0
imageio==2.34.0
ipython
jaxtyping==0.2.28
kmeans_gpu==0.0.5
matplotlib==3.7.2
mediapy==1.2.0
moviepy==1.0.3
numpy==1.25.0
omegaconf==2.1.1
open3d==0.18.0
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
Pillow
plyfile==1.0.3
point_cloud_utils==0.30.2
pyfqmr==0.2.0
pygltflib==1.16.2
PyMCubes==0.1.4
pymeshlab==2023.12
pyrender==0.1.45
robust-laplacian==0.2.7
roma==1.5.2.1
safetensors==0.3.3
scikit_learn==1.3.2
scipy
smplx==0.1.28
tensorboard
tqdm==4.65.0
trimesh==4.0.8
wandb[media]
warp_lang==0.10.1
xatlas==0.0.9